Multiple Tukey tests

Perform multiple Tukey tests using all levels of the categorical variables

tukey_various(dataset, kpi, categorical)

Arguments

dataset
Specify the dataframe

kpi
Specify the kpi variable (Must be included in the dataframe)

categorical
Specify a vector of categorical variables (Must be included in the dataframe)

Details

It returns a dataframe with multiple comparison of Tukey's range test on all levels of the categorical variables.

Anova returns the statistical significance of at least one level has significant difference in variance.

Examples

library(tidyverse)

t <- tukey_various(iris, "Sepal.Length", c("Species"))
t

#>                           Vars  diff       lwr       upr        p adj Variable
#> Species.2     virginica-setosa 1.582 1.3382273 1.8257727 2.997602e-15  Species
#> Species.1    versicolor-setosa 0.930 0.6862273 1.1737727 3.386180e-14  Species
#> Species.3 virginica-versicolor 0.652 0.4082273 0.8957727 8.287558e-09  Species